+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.19.0 (DEV)
 
+- Add: Spec Ports DB dialog now allows to edit entire DB as text.
 - Fix: Some UI texts correction in progress. Thanks to Lana Snik-Snak!
 - Fix: H/V/X Ports when set as Special wasn't marked Special in "Copy as Image".
 - Enh: H/V/X Ports in Toolbar are halved too.